Added a Book class Seperated Book object in another class
